requirements.txt
# Core ML
torch>=2.5.1
torchvision>=0.20.1

# Image processing
Pillow>=11.1.0
numpy>=2.0.1
scipy>=1.15.3

# CLIP dependencies
ftfy>=6.3.1
regex>=2026.1.15
tqdm>=4.67.3

# Utilities
packaging>=25.0